requirements.txt
librosa==0.8.0
numpy
safe_gpu
scipy==1.5.3
soundfile==0.10.3.post1
tensorboard
yamlargparse==1.31.1
pyannote.database